PP-11682 Refactor connector client code
- This PR is moving code around and deleting temp files.
- We introduced a `connector-axios.client` so that we could
  slowly move all connector related code to use the new Axios
  base client.
- All changes to connector code have now been completed.
- So we can now remove the temp `connector-axios.client` and
  transfer the changes to the `connector.client` instead.
    -2yy Take the contents of `connector-axios.client` and update the
    `connector.client`.
  - Update all references from `connector-axios.client` to
     `connector.client`.
  - Take the contents of `connector-axios.client` and update the
    `connector.client`.
  - Update the `connector.client` tests with the
    code from `connector-axios.client` tests.
  - Remove the `connector-axios.client` and tests.
