+++ app/controllers/charge.controller.js
@@ -28,7 +28,7 @@ const { countries } = require('../services/countries')
 const { commonTypos } = require('../utils/email-tools')
 const { withAnalyticsError, withAnalytics } = require('../utils/analytics')
 const { getMerchantId } = require('../utils/google-pay-merchant-id-selector')
-const connectorClient = require('../services/clients/connector-axios.client')
+const connectorClient = require('../services/clients/connector.client')
 const cookies = require('../utils/cookies')
 const { getGooglePayMethodData, googlePayDetails } = require('../utils/google-pay-check-request')
 const supportedNetworksFormattedByProvider = require('../assets/javascripts/browsered/web-payments/format-card-types')
