+++ app/models/charge.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const logger = require('../utils/logger')(__filename)
-const connectorClient = require('../services/clients/connector-axios.client')
+const connectorClient = require('../services/clients/connector.client')
 const State = require('../../config/state.js')
 const StateModel = require('../../config/state')
