+++ app/services/clients/connector.client.js
@@ -1,7 +1,8 @@
 'use strict'
 
 const logger = require('../../utils/logger')(__filename)
-const baseClient = require('./base.client/base.client')
+const { Client } = require('@govuk-pay/pay-js-commons/lib/utils/axios-base-client/axios-base-client')
+const { configureClient } = require('./base/config')
 const requestLogger = require('../../utils/request-logger')
 const { getCounter } = require('../../metrics/graphite-reporter')
 
@@ -18,8 +19,9 @@ const TOKEN_USED_PATH = '/v1/frontend/tokens/{chargeTokenId}/used'
 const CARD_CHARGE_PATH = '/v1/frontend/charges/{chargeId}'
 const WORLDPAY_3DS_FLEX_JWT_PATH = '/v1/frontend/charges/{chargeId}/worldpay/3ds-flex/ddc'
 
+const client = new Client(SERVICE_NAME)
+
 let baseUrl
-let correlationId
 
 /** @private */
 const _getFindChargeUrlFor = chargeId => baseUrl + CARD_CHARGE_PATH.replace('{chargeId}', chargeId)
@@ -60,7 +62,7 @@ const _getPatchUrlFor = chargeId => baseUrl + CARD_CHARGE_PATH.replace('{chargeI
 const _getWorldpay3dsFlexUrlFor = chargeId => baseUrl + WORLDPAY_3DS_FLEX_JWT_PATH.replace('{chargeId}', chargeId)
 
 /** @private */
-function _putConnector (url, payload, description, loggingFields = {}, callingFunctionName) {
+async function _putConnector (url, payload, description, loggingFields = {}, callingFunctionName) {
   const startTime = new Date()
   const context = {
     ...loggingFields,
@@ -70,37 +72,37 @@ function _putConnector (url, payload, description, loggingFields = {}, callingFu
     service: SERVICE_NAME
   }
   requestLogger.logRequestStart(context, loggingFields)
-  return baseClient
-    .put(url, { payload, correlationId })
-    .then(response => {
-      logger.info('PUT to %s ended - total time %dms', url, new Date() - startTime, loggingFields)
-      incrementStatusCodeCounter(callingFunctionName, response.statusCode)
-      if (response.statusCode > 499 && response.statusCode < 600) {
-        logger.error(`Error communicating with ${url}`, {
-          ...loggingFields,
-          service: 'connector',
-          method: 'PUT',
-          status_code: response.statusCode,
-          url: url
-        })
-      }
-      return response
-    }).catch(err => {
-      logger.info('PUT to %s ended - total time %dms', url, new Date() - startTime, loggingFields)
-      logger.error('Calling connector threw exception', {
+  try {
+    configureClient(client, url)
+    const response = await client.put(url, payload, description )
+    logger.info('PUT to %s ended - total time %dms', url, new Date() - startTime, loggingFields)
+    incrementStatusCodeCounter(callingFunctionName, response.status)
+    if (response.status > 499 && response.status < 600) {
+      logger.error(`Error communicating with ${url}`, {
         ...loggingFields,
         service: 'connector',
         method: 'PUT',
-        url: url,
-        error: err
+        status_code: response.status,
+        url: url
       })
-      incrementStatusCodeCounter(callingFunctionName, 'error')
-      throw err
+    }
+    return response
+  } catch (err) {
+    logger.info('PUT to %s ended - total time %dms', url, new Date() - startTime, loggingFields)
+    logger.error('Calling connector threw exception', {
+      ...loggingFields,
+      service: 'connector',
+      method: 'PUT',
+      url: url,
+      error: err
     })
+    incrementStatusCodeCounter(callingFunctionName, 'error')
+    throw err
+  }
 }
 
 /** @private */
-function _postConnector (url, payload, description, loggingFields = {}, callingFunctionName) {
+async function _postConnector (url, payload, description, loggingFields = {}, callingFunctionName) {
   const startTime = new Date()
   const context = {
     url: url,
@@ -109,23 +111,23 @@ function _postConnector (url, payload, description, loggingFields = {}, callingF
     service: SERVICE_NAME
   }
   requestLogger.logRequestStart(context, loggingFields)
-  return baseClient.post(
-    url,
-    { payload, correlationId }
-  ).then(response => {
+  try {
+    configureClient(client, url)
+
+    const response = await client.post(url, payload, description)
     logger.info('POST to %s ended - total time %dms', url, new Date() - startTime, loggingFields)
-    incrementStatusCodeCounter(callingFunctionName, response.statusCode)
-    if (response.statusCode > 499 && response.statusCode < 600) {
+    incrementStatusCodeCounter(callingFunctionName, response.status)
+    if (response.status > 499 && response.status < 600) {
       logger.error(`Error communicating with ${url}`, {
         ...loggingFields,
         service: 'connector',
         method: 'POST',
-        status_code: response.statusCode,
+        status_code: response.status,
         url: url
       })
     }
     return response
-  }).catch(err => {
+  } catch (err) {
     logger.info('POST to %s ended - total time %dms', url, new Date() - startTime, loggingFields)
     logger.error('Calling connector threw exception', {
       ...loggingFields,
@@ -136,11 +138,11 @@ function _postConnector (url, payload, description, loggingFields = {}, callingF
     })
     incrementStatusCodeCounter(callingFunctionName, 'error')
     throw err
-  })
+  }
 }
 
 /** @private */
-function _patchConnector (url, payload, description, loggingFields = {}, callingFunctionName) {
+async function _patchConnector (url, payload, description, loggingFields = {}, callingFunctionName) {
   const startTime = new Date()
   const context = {
     url: url,
@@ -149,23 +151,23 @@ function _patchConnector (url, payload, description, loggingFields = {}, calling
     service: SERVICE_NAME
   }
   requestLogger.logRequestStart(context, loggingFields)
-  return baseClient.patch(
-    url,
-    { payload, correlationId }
-  ).then(response => {
+  try {
+    configureClient(client, url)
+
+    const response = await client.patch(url, payload, description)
     logger.info('PATCH to %s ended - total time %dms', url, new Date() - startTime, loggingFields)
-    incrementStatusCodeCounter(callingFunctionName, response.statusCode)
-    if (response.statusCode > 499 && response.statusCode < 600) {
+    incrementStatusCodeCounter(callingFunctionName, response.status)
+    if (response.status > 499 && response.status < 600) {
       logger.error(`Error communicating with ${url}`, {
         ...loggingFields,
         service: 'connector',
         method: 'PATCH',
-        status_code: response.statusCode,
+        status_code: response.status,
         url: url
       })
     }
     return response
-  }).catch(err => {
+  } catch (err) {
     logger.info('PATCH %s to %s ended - total time %dms', url, new Date() - startTime, loggingFields)
     logger.error('Calling connector threw exception', {
       ...loggingFields,
@@ -176,11 +178,11 @@ function _patchConnector (url, payload, description, loggingFields = {}, calling
     })
     incrementStatusCodeCounter(callingFunctionName, 'error')
     throw err
-  })
+  }
 }
 
 /** @private */
-function _getConnector (url, description, loggingFields = {}, callingFunctionName) {
+async function _getConnector (url, description, loggingFields = {}, callingFunctionName) {
   const startTime = new Date()
   const context = {
     url: url,
@@ -189,32 +191,34 @@ function _getConnector (url, description, loggingFields = {}, callingFunctionNam
     service: SERVICE_NAME
   }
   requestLogger.logRequestStart(context, loggingFields)
-  return baseClient.get(url, { correlationId })
-    .then(response => {
-      logger.info('GET to %s ended - total time %dms', url, new Date() - startTime, loggingFields)
-      incrementStatusCodeCounter(callingFunctionName, response.statusCode)
-      if (response.statusCode !== 200) {
-        logger.error(`Error communicating with ${url}`, {
-          ...loggingFields,
-          service: 'connector',
-          method: 'GET',
-          status_code: response.statusCode,
-          url: url
-        })
-      }
-      return response
-    }).catch(err => {
-      logger.info('GET to %s ended - total time %dms', url, new Date() - startTime, loggingFields)
-      logger.error('Calling connector threw exception', {
+  try {
+    configureClient(client, url)
+
+    const response = await client.get(`${url}`, description)
+    logger.info('GET to %s ended - total time %dms', url, new Date() - startTime, loggingFields)
+    incrementStatusCodeCounter(callingFunctionName, response.status)
+    if (response.status !== 200) {
+      logger.error(`Error communicating with ${url}`, {
         ...loggingFields,
         service: 'connector',
         method: 'GET',
-        url: url,
-        error: err
+        status_code: response.status,
+        url: url
       })
-      incrementStatusCodeCounter(callingFunctionName, 'error')
-      throw err
+    }
+    return response
+  } catch (err) {
+    logger.info('GET to %s ended - total time %dms', url, new Date() - startTime, loggingFields)
+    logger.error('Calling connector threw exception', {
+      ...loggingFields,
+      service: 'connector',
+      method: 'GET',
+      url: url,
+      error: err
     })
+    incrementStatusCodeCounter(callingFunctionName, 'error')
+    throw err
+  }
 }
 
 const incrementStatusCodeCounter = (callingFunctionName, statusCode) => {
@@ -282,7 +286,6 @@ const markTokenAsUsed = (chargeOptions, loggingFields = {}) => {
 
 module.exports = function (clientOptions = {}) {
   baseUrl = clientOptions.baseUrl || process.env.CONNECTOR_HOST
-  correlationId = clientOptions.correlationId || ''
   return {
     chargeAuth,
     chargeAuthWithWallet,
