+++ app/services/worldpay-3ds-flex.service.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const connectorClient = require('./clients/connector-axios.client')
+const connectorClient = require('./clients/connector.client')
 
 const getDdcJwt = async function getDdcJwt (charge, correlationId, loggingFields = {}) {
   if (charge.paymentProvider === 'worldpay' &&
